spoke_server: disclaim SPOKEPORT global arg
the args=(SPOKEPORT) was giving
argument after * must be an iterable, not int
Found nothing wrong with it; it works if it has anything else,
filler data etc in the arguments.

No idea why it happened.
Since moving around things, SPOKEPORT is now actually reachable,
thus the args part can be removed instead.